requirements.txt
fastapi
uvicorn
celery
requests
sqlalchemy
python-dotenv
asyncpg
alembic
psycopg2-binary
httpx
apscheduler
pytest
pytest-asyncio